agent_analyst/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "blockchain-analyst-agent"
version = "0.1.0"
description = "Professional blockchain analyst agent with x402 micropayments"
dependencies = [
    "mcp>=1.0.0",
    "pydantic>=2.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-dotenv>=1.0.0",
    "x402>=0.1.0",
    "cdp-sdk>=0.1.0",
    "httpx>=0.25.0",
    "google-generativeai>=0.8.0"
]

[tool.hatch.build.targets.wheel]
packages = ["blockchain_analyst"]

[project.scripts]
blockchain-analyst-agent = "blockchain_analyst.x402_server:main"
